Adds a number of modifications to the calibration flow:
* Renamed script from calibration_pipeline.py to just calibrate.py
* Introduced a number of arguments to execute portiions of the pipeline and only for specified polygon positions
* Doesn't throw an error if the datasets/ dir already exists
* Defaults to 1 image per polygon position (vs. 5) so total default images = 13 vs. 65 for a faster flow.
* Adds more incremental logging messages
